Documents/Hackathons/meta/NeuralKarma copy/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
datasets==2.15.0
scikit-learn==1.3.2
joblib==1.3.2
pandas==2.1.4
numpy==1.26.2
jinja2==3.1.2
python-multipart==0.0.6
aiofiles==23.2.1
websockets==12.0
huggingface-hub==0.19.4
aiohttp==3.9.1
openai==1.3.9
python-dotenv==1.0.0
